Refactor plan validation
